feat: adicionada impedimento para deletar tarefa

+++ src/pages/api/tarefas/deletarTarefa.js
@@ -13,6 +13,15 @@ const handler = async (req, res) => {
             return res.status(400).json(defaultResponse('Preencha todos os dados para continuar'));
         }
 
+        const hasFiles = await db.query({ text: 'SELECT 1 FROM tarefa_arquivo WHERE id_tarefa = $1', values:[id] });
+
+        console.log(hasFiles.rowCount);
+        console.log(hasFiles.rows);
+
+        if(hasFiles.rowCount > 0){
+            return res.status(409).json(defaultResponse('Delete os arquivos antes de deletar as tarefas'));
+        }
+
         const sql = `
             DELETE FROM tarefa
             WHERE id = $1
